feat/refactor: implement functionality of list jobs by user id and improve name of some classes

+++ src/main/java/com/miguel/jobnest/application/usecases/usercode/DefaultResendVerificationCodeUseCase.java
@@ -4,21 +4,24 @@ import com.miguel.jobnest.application.abstractions.producer.MessageProducer;
 import com.miguel.jobnest.application.abstractions.providers.CodeProvider;
 import com.miguel.jobnest.application.abstractions.repositories.UserCodeRepository;
 import com.miguel.jobnest.application.abstractions.repositories.UserRepository;
-import com.miguel.jobnest.application.abstractions.usecases.usercode.SendPasswordResetCodeUseCase;
-import com.miguel.jobnest.application.usecases.usercode.inputs.SendPasswordResetCodeUseCaseInput;
+import com.miguel.jobnest.application.abstractions.usecases.usercode.ResendVerificationCodeUseCase;
+import com.miguel.jobnest.application.usecases.usercode.inputs.ResendVerificationCodeUseCaseInput;
 import com.miguel.jobnest.domain.entities.User;
 import com.miguel.jobnest.domain.entities.UserCode;
 import com.miguel.jobnest.domain.enums.UserCodeType;
+import com.miguel.jobnest.domain.enums.UserStatus;
 import com.miguel.jobnest.domain.events.UserCodeCreatedEvent;
+import com.miguel.jobnest.domain.exceptions.DomainException;
 import com.miguel.jobnest.domain.exceptions.NotFoundException;
 
+import java.util.Objects;
 import java.util.Optional;
 
-public class SendPasswordResetCodeUseCaseImpl implements SendPasswordResetCodeUseCase {
-    private final UserRepository userRepository;
+public class DefaultResendVerificationCodeUseCase implements ResendVerificationCodeUseCase {
     private final UserCodeRepository userCodeRepository;
-    private final CodeProvider codeProvider;
+    private final UserRepository userRepository;
     private final MessageProducer messageProducer;
+    private final CodeProvider codeProvider;
 
     private final static int CODE_LENGTH = 6;
     private final static String ALPHANUMERIC_CHARACTERS = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
@@ -26,21 +29,29 @@ public class SendPasswordResetCodeUseCaseImpl implements SendPasswordResetCodeUs
     private static final String USER_CODE_CREATED_EXCHANGE = "user.code.created.exchange";
     private static final String USER_CODE_CREATED_ROUTING_KEY = "user.code.created.routing.key";
 
-    public SendPasswordResetCodeUseCaseImpl(
-            UserRepository userRepository,
+    public DefaultResendVerificationCodeUseCase(
             UserCodeRepository userCodeRepository,
-            CodeProvider codeProvider,
-            MessageProducer messageProducer
+            UserRepository userRepository,
+            MessageProducer messageProducer,
+            CodeProvider codeProvider
     ) {
-        this.userRepository = userRepository;
         this.userCodeRepository = userCodeRepository;
-        this.codeProvider = codeProvider;
+        this.userRepository = userRepository;
         this.messageProducer = messageProducer;
+        this.codeProvider = codeProvider;
     }
 
     @Override
-    public void execute(SendPasswordResetCodeUseCaseInput input) {
+    public void execute(ResendVerificationCodeUseCaseInput input) {
         final User user = this.getUserByEmail(input.email());
+
+        if (Objects.equals(user.getUserStatus(), UserStatus.VERIFIED)) {
+            throw DomainException.with("User has already been verified", 409);
+        }
+
+        if (Objects.equals(user.getUserStatus(), UserStatus.DELETED)) {
+            throw DomainException.with("User has already been deleted", 409);
+        }
 
         this.getPreviousUserCodeByUserIdAndCodeType(user.getId()).ifPresent(userCode ->
                 this.deleteUserCodeById(userCode.getId())
@@ -48,7 +59,7 @@ public class SendPasswordResetCodeUseCaseImpl implements SendPasswordResetCodeUs
 
         final String codeGenerated = this.codeProvider.generateCode(CODE_LENGTH, ALPHANUMERIC_CHARACTERS);
 
-        final UserCode newUserCode = UserCode.newUserCode(user.getId(), codeGenerated, UserCodeType.PASSWORD_RESET);
+        final UserCode newUserCode = UserCode.newUserCode(user.getId(), codeGenerated, UserCodeType.USER_VERIFICATION);
 
         final UserCode savedUserCode = this.saveUserCode(newUserCode);
